+++ zingolib/src/test_framework.rs
@@ -1,63 +1,51 @@
-pub(crate) mod macros;
-
 use zcash_primitives::transaction::TxId;
 
-macro_rules! build_method {
-    ($name:ident, $localtype:ty) => {
-        pub fn $name(mut self, $name: $localtype) -> Self {
-            self.$name = Some($name);
-            self
-        }
-    };
-}
 use crate::wallet::notes::TransparentNote;
-pub struct TransparentNoteBuilder {
-    address: Option<String>,
-    txid: Option<TxId>,
-    output_index: Option<u64>,
-    script: Option<Vec<u8>>,
-    value: Option<u64>,
-    spent: Option<Option<(TxId, u32)>>,
-    unconfirmed_spent: Option<Option<(TxId, u32)>>,
-}
-#[allow(dead_code)] //TODO:  fix this gross hack that I tossed in to silence the language-analyzer false positive
-impl TransparentNoteBuilder {
-    pub fn new() -> Self {
-        Self::default()
-    }
-    // Methods to set each field
-    build_method!(address, String);
-    build_method!(txid, TxId);
-    build_method!(output_index, u64);
-    build_method!(script, Vec<u8>);
-    build_method!(value, u64);
-    build_method!(spent, Option<(TxId, u32)>);
-    build_method!(unconfirmed_spent, Option<(TxId, u32)>);
+pub(crate) mod macros;
+mod mocks;
 
-    // Build method
-    pub fn build(self) -> TransparentNote {
-        TransparentNote::from_parts(
-            self.address.unwrap(),
-            self.txid.unwrap(),
-            self.output_index.unwrap(),
-            self.script.unwrap(),
-            self.value.unwrap(),
-            self.spent.unwrap(),
-            self.unconfirmed_spent.unwrap(),
-        )
-    }
+#[allow(dead_code)]
+pub(crate) fn create_empty_txid_and_tnote() -> (zcash_primitives::transaction::TxId, TransparentNote)
+{
+    // A single transparent note makes is_incoming_trsaction true.
+    let txid = zcash_primitives::transaction::TxId::from_bytes([0u8; 32]);
+    (
+        txid,
+        mocks::TransparentNoteBuilder::new()
+            .address("t".to_string())
+            .spent(Some((txid, 3)))
+            .build(),
+    )
 }
-
-impl Default for TransparentNoteBuilder {
-    fn default() -> Self {
-        TransparentNoteBuilder {
-            address: Some("default_address".to_string()),
-            txid: Some(TxId::from_bytes([0u8; 32])),
-            output_index: Some(0),
-            script: Some(vec![]),
-            value: Some(0),
-            spent: Some(None),
-            unconfirmed_spent: Some(None),
-        }
+#[allow(dead_code)]
+pub(crate) fn create_transaction_record_with_one_tnote(
+    txid: TxId,
+    transparent_note: TransparentNote,
+) -> crate::wallet::transaction_record::TransactionRecord {
+    // A single transparent note makes is_incoming_trsaction true.
+    let mut transaction_record = crate::wallet::transaction_record::TransactionRecord::new(
+        zingo_status::confirmation_status::ConfirmationStatus::Confirmed(
+            zcash_primitives::consensus::BlockHeight::from_u32(5),
+        ),
+        1705077003,
+        &txid,
+    );
+    transaction_record.transparent_notes.push(transparent_note);
+    transaction_record
+}
+#[allow(dead_code)]
+pub(crate) fn default_trecord_with_one_tnote(
+) -> crate::wallet::transaction_record::TransactionRecord {
+    let (txid, transparent_note) = create_empty_txid_and_tnote();
+    create_transaction_record_with_one_tnote(txid, transparent_note)
+}
+#[allow(dead_code)]
+pub(crate) fn create_note_record_id() -> crate::wallet::notes::NoteRecordIdentifier {
+    let (txid, _tnote) = create_empty_txid_and_tnote();
+    let index = 5u32;
+    crate::wallet::notes::NoteRecordIdentifier {
+        txid,
+        pool: zcash_client_backend::PoolType::Transparent,
+        index,
     }
 }